refactor: todoType 내용 todo.type.ts로 통합
