+++ types/todo.type.ts
@@ -1,14 +1,21 @@
 import type { GoalType } from './goal.type';
 
+export type TodoListType = {
+  lastSeenId: number;
+  totalCount: number;
+  todos: TodoType[];
+};
+
 export type TodoType = {
   goal: Pick<GoalType, 'id' | 'title'>;
-  noteId: number;
-  done: boolean;
+  noteId: number | null;
+  done: boolean | null;
   linkUrl: string | null;
   filePath: string | null;
   title: string;
   id: number;
   userId: number;
-  createdAt: string;
-  updatedAt: string;
+  createdAt: Date;
+  updatedAt: Date;
+  priority: number;
 };
